Firmware: Move LED code into own file
